Updated app Login now on separate page Signup not working yet Every page in the app have the directive for sidebar, topbar and footer
