+++ src/assets/scripts/ads-details/ads-details.html
@@ -1,65 +1,71 @@
-<main class='main-content bgc-grey-300'>
-  <div id='mainContent row'>
+<nw-sidebar></nw-sidebar>
+<div class="page-container">
+  <nw-topbar></nw-topbar>
+  <main class='main-content bgc-grey-300'>
+    <div id='mainContent row'>
 
-    <div class="content-title mT-60">
-      <div class="ml-1 mb-3">
-        <h1>Ad's Details</h1>
+      <div class="content-title mT-60">
+        <div class="ml-1 mb-3">
+          <h1>Ad's Details</h1>
 
+        </div>
       </div>
+
+      <div class=" gap-30 masonry pos-r ">
+        <!-- #content container ===== -->
+        <div class="masonry-sizer col-12 "></div>
+
+        <!-- ### Job Details Card ### -->
+        <div class="masonry-item  w-100">
+          <div class="card">
+            <div class="row">
+              <div class="col-2 mt-4 ">
+                <img class="img-fluid ml-4 mb-4 companies-update-img avatar rounded-circle" id="ad-details-avatar" alt="company"
+                  ng-src={{avatar.element.src}} />
+              </div>
+              <div class="col-10 card-body">
+                <h2 class="card-title mt-4 mr-4">{{ ad.title }}</h2>
+                <h6 class="card-title mr-4">{{ ad.companyname }}</h6>
+              </div>
+            </div>
+          </div>
+          <div class="card">
+            <div class="card-body">
+              {{ ad.description }}
+            </div>
+          </div>
+        </div>
+
+        <!-- ### Other Info Card ### -->
+        <div class="masonry-item  w-100">
+          <div class="card">
+            <div class="card-body">
+              <h5 class="card-title mt-4 ml-4 mr-4">Location:</h5>
+              <p class="card-text ml-4 mr-4">{{ad.location}}</p>
+              <h5 class="card-title mt-4 ml-4 mr-4">Application close:</h5>
+              <p class="card-text ml-4 mr-4">{{ad.expiration-date}}</p>
+
+              <h5 class="card-title mt-4 ml-4 mr-4">Skills:</h5>
+              <div class="ml-4 mb-3" ng-repeat="skill in ad.skills">
+                <span class="mr-2 mb-2 btn btn-skills">{{ skill }}</span>
+              </div>
+              <p class="card-text  ml-4 my-4"></p><small class="text-muted">{{ daysAgo(ad.timestamp) }}</small></p>
+            </div>
+          </div>
+        </div>
+
+        <!-- ### Buttons ### -->
+        <div class="masonry-item  w-50 m-auto btn-centered-container visible-user" ng-show="profile.role =='user'">
+          <button type="button" href="#" class="col-5 mr-1 btn btn-gradient btn-centered">Apply Now</button>
+          <button type="button" href="#" class="col-5 ml-1 btn btn-border  btn-centered">Save</button>
+
+        </div>
+      </div>
+
+
     </div>
+    <!--content row-->
 
-    <div class=" gap-30 masonry pos-r ">
-      <!-- #content container ===== -->
-      <div class="masonry-sizer col-12 "></div>
-
-      <!-- ### Job Details Card ### -->
-      <div class="masonry-item  w-100">
-        <div class="card">
-          <div class="row">
-            <div class="col-2 mt-4 ">
-              <img class="img-fluid ml-4 mb-4 companies-update-img avatar rounded-circle" id="ad-details-avatar" alt="company" ng-src={{avatar.element.src}} />
-            </div>
-            <div class="col-10 card-body">
-              <h2 class="card-title mt-4 mr-4">{{ ad.title }}</h2>
-              <h6 class="card-title mr-4">{{ ad.companyname }}</h6>
-            </div>
-          </div>
-        </div>
-        <div class="card">
-          <div class="card-body">
-            {{ ad.description }}
-          </div>
-        </div>
-      </div>
-
-      <!-- ### Other Info Card ### -->
-      <div class="masonry-item  w-100">
-        <div class="card">
-          <div class="card-body">
-            <h5 class="card-title mt-4 ml-4 mr-4">Location:</h5>
-            <p class="card-text ml-4 mr-4">{{ad.location}}</p>
-            <h5 class="card-title mt-4 ml-4 mr-4">Application close:</h5>
-            <p class="card-text ml-4 mr-4">{{ad.expiration-date}}</p>
-
-            <h5 class="card-title mt-4 ml-4 mr-4">Skills:</h5>
-            <div class="ml-4 mb-3" ng-repeat="skill in ad.skills">
-              <span  class="mr-2 mb-2 btn btn-skills">{{ skill }}</span>
-            </div>
-            <p class="card-text  ml-4 my-4"></p><small class="text-muted">{{ daysAgo(ad.timestamp) }}</small></p>
-          </div>
-        </div>
-      </div>
-
-      <!-- ### Buttons ### -->
-      <div class="masonry-item  w-50 m-auto btn-centered-container visible-user" ng-show="profile.role =='user'">
-        <button type="button" href="#" class="col-5 mr-1 btn btn-gradient btn-centered">Apply Now</button>
-        <button type="button" href="#" class="col-5 ml-1 btn btn-border  btn-centered">Save</button>
-
-      </div>
-    </div>
-
-
-  </div>
-  <!--content row-->
-
-</main>
+  </main>
+  <nw-footer></nw-footer>
+</div>
