+++ src/assets/scripts/companies/companies.html
@@ -1,49 +1,48 @@
-<main class='main-content'>
-  <div id='mainContent row'>
-    <div class="row content-title mT-60">
-      <h1 class="col-8 ml-1 mb-4">Companies</h1>
-    </div>
+<nw-sidebar></nw-sidebar>
+<div class="page-container">
+  <nw-topbar></nw-topbar>
+  <main class='main-content'>
+    <div id='mainContent row'>
+      <div class="row content-title mT-60">
+        <h1 class="col-8 ml-1 mb-4">Companies</h1>
+      </div>
 
-    <div class=" gap-30 masonry pos-r">
-      <!-- #content container ===== -->
-      <div class="masonry-sizer col-md-6"></div>
+      <div class=" gap-30 masonry pos-r">
+        <!-- #content container ===== -->
+        <div class="masonry-sizer col-md-6"></div>
 
-      <div class="masonry-item w-50" ng-repeat="company in companies">
-        <!-- #first user ========= -->
-        <div class="card item-box">
-          <div class="user-data-title container">
-            <img class="user-card-img rounded-circle" id="company{{$index}}" ng-src="{{avatar($index)}}" alt="">
-            <div class="user-card-col">
-              <div class="user-card-title">{{company.name}}</div>
-              <div class=" text-size-11">Victoria</div>
-              <div class=" text-size-11">{{companies.email}}</div>
+        <div class="masonry-item w-50" ng-repeat="company in companies">
+          <!-- #first user ========= -->
+          <div class="card item-box">
+            <div class="user-data-title container">
+              <img class="user-card-img rounded-circle" id="company{{$index}}" ng-src="{{avatar($index)}}" alt="">
+              <div class="user-card-col">
+                <div class="user-card-title">{{company.name}}</div>
+                <div class=" text-size-11">Victoria</div>
+                <div class=" text-size-11">{{companies.email}}</div>
 
-              <a href="#!company/{{company._id}}" id="user-action-fill">view company profile</a>
-              <a href="#" id="user-action-border" data-toggle="tooltip" data-placement="bottom" title="Company will be unable to log in to WISP">ban
-                company</a>
+                <a href="#!company/{{company._id}}" id="user-action-fill">view company profile</a>
+                <a href="#" id="user-action-border" data-toggle="tooltip" data-placement="bottom" title="Company will be unable to log in to WISP">ban
+                  company</a>
+              </div>
             </div>
-          </div>
-          <div class="card-body p-0">
-            <div class="card-body-users row">
-              <div class="col-4"></div>
-              <div class="users-icons col-8">
-                <a href=""><i class="fa fa-envelope-o"></i></a>
-                <a href=""><i class="fa fa-linkedin"></i></a>
-                <a href=""><i class="fa fa-facebook"></i></a>
-                <a href=""><i class="fa fa-instagram"></i></a>
-                <a href=""><i class="fa fa-github"></i></a>
+            <div class="card-body p-0">
+              <div class="card-body-users row">
+                <div class="col-4"></div>
+                <div class="users-icons col-8">
+                  <a href=""><i class="fa fa-envelope-o"></i></a>
+                  <a href=""><i class="fa fa-linkedin"></i></a>
+                  <a href=""><i class="fa fa-facebook"></i></a>
+                  <a href=""><i class="fa fa-instagram"></i></a>
+                  <a href=""><i class="fa fa-github"></i></a>
+                </div>
               </div>
             </div>
           </div>
         </div>
       </div>
-
-      
     </div>
-
-
-
-  </div>
-  <!--content container-->
-
-</main>
+    <!--content container-->
+  </main>
+  <nw-footer></nw-footer>
+</div>
