+++ src/assets/scripts/history/history.html
@@ -1,72 +1,77 @@
-<main class='main-content bgc-grey-300'>
-  <div id='mainContent row'>
-    <div class="row content-title mT-60">
-      <h1 class="col-8 ml-1 mb-3">
-        History
-      </h1>
+<nw-sidebar></nw-sidebar>
+<div class="page-container">
+  <nw-topbar></nw-topbar>
+  <main class='main-content bgc-grey-300'>
+    <div id='mainContent row'>
+      <div class="row content-title mT-60">
+        <h1 class="col-8 ml-1 mb-3">
+          History
+        </h1>
+      </div>
+      <!-- ### Table ### -->
+      <table class="table table-hover">
+        <thead class="thead-purple-border">
+          <tr>
+            <th scope="col">#</th>
+            <th scope="col">Position</th>
+            <th scope="col">Company</th>
+            <th scope="col">Industry</th>
+            <th scope="col">Submission Date</th>
+            <th scope="col">Application Status</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr>
+            <th scope="row">1</th>
+            <td style="text-decoration:underline"><a href="ads_1.html">Graphic Design Intern</a></td>
+            <td>Sierra Delta Creative</td>
+            <td>Design</td>
+            <td>07 Jan 2019</td>
+            <td>
+              <div class="btn btn-outline-success btn-sm" data-toggle="tooltip" data-placement="right" title="Application is still open">
+                ON
+              </div>
+            </td>
+          </tr>
+          <tr>
+            <th scope="row">2</th>
+            <td style="text-decoration:underline"><a href="ads_1.html">Barista Intern</a></td>
+            <td>Brunetti</td>
+            <td>Hospitality</td>
+            <td>21 Jan 2019</td>
+            <td>
+              <div class="btn btn-outline-secondary btn-sm" data-toggle="tooltip" data-placement="right" title="Application has expired">
+                OFF
+              </div>
+            </td>
+          </tr>
+          <tr>
+            <th scope="row">3</th>
+            <td style="text-decoration:underline"><a href="ads_1.html">Web Developer Intern</a></td>
+            <td>Agile Talent</td>
+            <td>IT / Software Engineering</td>
+            <td>23 Jan 2019</td>
+            <td>
+              <div class="btn btn-outline-info btn-sm" data-toggle="tooltip" data-placement="right" title="Application is closed">
+                CLOSED
+              </div>
+            </td>
+          </tr>
+          <tr>
+            <th scope="row">4</th>
+            <td style="text-decoration:underline"><a href="ads_1.html">Kitchen Hand Intern</a></td>
+            <td>Yarra River Grill</td>
+            <td>Hospitality</td>
+            <td>25 Jan 2019</td>
+            <td>
+              <div class="btn btn-outline-danger btn-sm" data-toggle="tooltip" data-placement="right" title="This job listing is removed">
+                CANCELLED
+              </div>
+            </td>
+          </tr>
+        </tbody>
+      </table>
     </div>
-    <!-- ### Table ### -->
-    <table class="table table-hover">
-      <thead class="thead-purple-border">
-        <tr>
-          <th scope="col">#</th>
-          <th scope="col">Position</th>
-          <th scope="col">Company</th>
-          <th scope="col">Industry</th>
-          <th scope="col">Submission Date</th>
-          <th scope="col">Application Status</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr>
-          <th scope="row">1</th>
-          <td style="text-decoration:underline"><a href="ads_1.html">Graphic Design Intern</a></td>
-          <td>Sierra Delta Creative</td>
-          <td>Design</td>
-          <td>07 Jan 2019</td>
-          <td>
-            <div class="btn btn-outline-success btn-sm" data-toggle="tooltip" data-placement="right" title="Application is still open">
-              ON
-            </div>
-          </td>
-        </tr>
-        <tr>
-          <th scope="row">2</th>
-          <td style="text-decoration:underline"><a href="ads_1.html">Barista Intern</a></td>
-          <td>Brunetti</td>
-          <td>Hospitality</td>
-          <td>21 Jan 2019</td>
-          <td>
-            <div class="btn btn-outline-secondary btn-sm" data-toggle="tooltip" data-placement="right" title="Application has expired">
-              OFF
-            </div>
-          </td>
-        </tr>
-        <tr>
-          <th scope="row">3</th>
-          <td style="text-decoration:underline"><a href="ads_1.html">Web Developer Intern</a></td>
-          <td>Agile Talent</td>
-          <td>IT / Software Engineering</td>
-          <td>23 Jan 2019</td>
-          <td>
-            <div class="btn btn-outline-info btn-sm" data-toggle="tooltip" data-placement="right" title="Application is closed">
-              CLOSED
-            </div>
-          </td>
-        </tr>
-        <tr>
-          <th scope="row">4</th>
-          <td style="text-decoration:underline"><a href="ads_1.html">Kitchen Hand Intern</a></td>
-          <td>Yarra River Grill</td>
-          <td>Hospitality</td>
-          <td>25 Jan 2019</td>
-          <td>
-            <div class="btn btn-outline-danger btn-sm" data-toggle="tooltip" data-placement="right" title="This job listing is removed">
-              CANCELLED
-            </div>
-          </td>
-        </tr>
-      </tbody>
-    </table>
-  </div>
-</main>
+  </main>
+  <nw-footer></nw-footer>
+</div>
