+++ src/assets/scripts/login/index.js
@@ -1,0 +1,8 @@
+import * as angular from 'angular';
+import ngCookies from 'angular-cookies';
+import LoginController from './login.controller';
+import LoginService from './login.service';
+
+angular.module('nwLogin', [ngCookies])
+    .controller('LoginController', LoginController)
+    .factory('LoginService', LoginService);
