+++ src/assets/scripts/referees/referees.html
@@ -1,47 +1,49 @@
-<main class='main-content'>
-  <div id='mainContent row'>
-    <div class="row content-title mT-60">
-      <h1 class="col-8 ml-1 mb-4">Referees</h1>
-    </div>
+<nw-sidebar></nw-sidebar>
+<div class="page-container">
+  <nw-topbar></nw-topbar>
 
-    <div class=" gap-30 masonry pos-r">
-      <!-- #content container ===== -->
-      <div class="masonry-sizer col-md-6"></div>
+  <main class='main-content'>
+    <div id='mainContent row'>
+      <div class="row content-title mT-60">
+        <h1 class="col-8 ml-1 mb-4">Referees</h1>
+      </div>
 
-      <div class="masonry-item  w-50" ng-repeat="referee in referees">
-        <!-- #first user ========= -->
-        <div class="card item-box">
-          <div class="user-data-title container">
-            <img class="user-card-img  rounded-circle" src="https://www.monmouth.edu/university-advancement/wp-content/uploads/sites/237/2018/08/person-placeholder.jpg"
-              alt="">
-            <div class="user-card-col">
-              <div class="user-card-title">{{referee.name}}</div>
-              <div class=" text-size-11">Victoria</div>
-              <div class=" text-size-11">{{referee.email}}</div>
-              <a href="#" id="user-action-fill">{{referee.bio}}</a>
-              <a href="#" id="user-action-border" data-toggle="tooltip" data-placement="bottom" title="Referee will be unable to log in to WISP">ban
-                referee</a>
+      <div class=" gap-30 masonry pos-r">
+        <!-- #content container ===== -->
+        <div class="masonry-sizer col-md-6"></div>
+
+        <div class="masonry-item  w-50" ng-repeat="referee in referees">
+          <!-- #first user ========= -->
+          <div class="card item-box">
+            <div class="user-data-title container">
+              <img class="user-card-img  rounded-circle" src="https://www.monmouth.edu/university-advancement/wp-content/uploads/sites/237/2018/08/person-placeholder.jpg"
+                alt="">
+              <div class="user-card-col">
+                <div class="user-card-title">{{referee.name}}</div>
+                <div class=" text-size-11">Victoria</div>
+                <div class=" text-size-11">{{referee.email}}</div>
+                <a href="#" id="user-action-fill">{{referee.bio}}</a>
+                <a href="#" id="user-action-border" data-toggle="tooltip" data-placement="bottom" title="Referee will be unable to log in to WISP">ban
+                  referee</a>
+              </div>
             </div>
-          </div>
-          <div class="card-body p-0">
-            <div class="card-body-users row">
-              <div class="col-4"></div>
-              <div class="users-icons col-8">
-                <a href=""><i class="fa fa-envelope-o"></i></a>
-                <a href=""><i class="fa fa-linkedin"></i></a>
-                <a href=""><i class="fa fa-facebook"></i></a>
-                <a href=""><i class="fa fa-instagram"></i></a>
-                <a href=""><i class="fa fa-github"></i></a>
+            <div class="card-body p-0">
+              <div class="card-body-users row">
+                <div class="col-4"></div>
+                <div class="users-icons col-8">
+                  <a href=""><i class="fa fa-envelope-o"></i></a>
+                  <a href=""><i class="fa fa-linkedin"></i></a>
+                  <a href=""><i class="fa fa-facebook"></i></a>
+                  <a href=""><i class="fa fa-instagram"></i></a>
+                  <a href=""><i class="fa fa-github"></i></a>
+                </div>
               </div>
             </div>
           </div>
         </div>
       </div>
-
     </div>
-
-
-  </div>
-  <!--content container-->
-
-</main>
+    <!--content container-->
+  </main>
+  <nw-footer></nw-footer>
+</div>
