+++ src/assets/scripts/users/users.html
@@ -1,45 +1,48 @@
-<main class='main-content' ng-controller="UsersController" id="users">
-  <div id='mainContent row'>
-    <div class="row content-title mT-60">
-      <h1 class="col-8 ml-1 mb-4">Users</h1>
-    </div>
+<nw-sidebar></nw-sidebar>
+<div class="page-container">
+  <nw-topbar></nw-topbar>
+  <main class='main-content' ng-controller="UsersController" id="users">
+    <div id='mainContent row'>
+      <div class="row content-title mT-60">
+        <h1 class="col-8 ml-1 mb-4">Users</h1>
+      </div>
 
-    <div class=" gap-30 masonry pos-r">
-      <!-- #content container ===== -->
-      <div class="masonry-sizer col-md-6"></div>
+      <div class=" gap-30 masonry pos-r">
+        <!-- #content container ===== -->
+        <div class="masonry-sizer col-md-6"></div>
 
-      <div class="masonry-item w-50" ng-repeat="u in users">
-        <!-- #first user ========= -->
-        <div class="card item-box">
-          <div class="user-data-title container">
-            <img id="user{{$index}}" class="user-card-img rounded-circle" ng-src="{{avatar($index)}}" alt="" />
-            <div class="user-card-col">
-              <div class="user-card-title">{{u.personal.first_name}} {{u.personal.last_name}}</div>
-              <div class=" text-size-11">{{u.personal.city}}</div>
-              <div class=" text-size-11">{{u.personal.phone}}</div>
-              <a href="#!users/{{u._id}}" id="user-action-fill">view user profile</a>
-              <a href="#" class="visible-nomit visible-referee visible-admin" ng-show="user.role =='nomit' || user.role =='referee' || user.role =='admin'" id="user-action-border" data-toggle="tooltip" data-placement="bottom" title="User will be unable to log in to WISP">
-                ban user</a>
+        <div class="masonry-item w-50" ng-repeat="u in users">
+          <!-- #first user ========= -->
+          <div class="card item-box">
+            <div class="user-data-title container">
+              <img id="user{{$index}}" class="user-card-img rounded-circle" ng-src="{{avatar($index)}}" alt="" />
+              <div class="user-card-col">
+                <div class="user-card-title">{{u.personal.first_name}} {{u.personal.last_name}}</div>
+                <div class=" text-size-11">{{u.personal.city}}</div>
+                <div class=" text-size-11">{{u.personal.phone}}</div>
+                <a href="#!users/{{u._id}}" id="user-action-fill">view user profile</a>
+                <a href="#" class="visible-nomit visible-referee visible-admin" ng-show="user.role =='nomit' || user.role =='referee' || user.role =='admin'"
+                  id="user-action-border" data-toggle="tooltip" data-placement="bottom" title="User will be unable to log in to WISP">
+                  ban user</a>
+              </div>
             </div>
-          </div>
-          <div class="card-body p-0">
-            <div class="card-body-users row">
-              <div class="col-4"></div>
-              <div class="users-icons col-8">
-                <a href=""><i class="fa fa-envelope-o"></i></a>
-                <a href=""><i class="fa fa-linkedin"></i></a>
-                <a href=""><i class="fa fa-facebook"></i></a>
-                <a href=""><i class="fa fa-instagram"></i></a>
-                <a href=""><i class="fa fa-github"></i></a>
+            <div class="card-body p-0">
+              <div class="card-body-users row">
+                <div class="col-4"></div>
+                <div class="users-icons col-8">
+                  <a href=""><i class="fa fa-envelope-o"></i></a>
+                  <a href=""><i class="fa fa-linkedin"></i></a>
+                  <a href=""><i class="fa fa-facebook"></i></a>
+                  <a href=""><i class="fa fa-instagram"></i></a>
+                  <a href=""><i class="fa fa-github"></i></a>
+                </div>
               </div>
             </div>
           </div>
         </div>
       </div>
 
-
-    </div>
-
-  <!--content container-->
-
-</main>
+      <!--content container-->
+  </main>
+  <nw-footer></nw-footer>
+</div>
