+++ src/index.html
@@ -71,29 +71,7 @@
     });
   </script>
 
-  <!-- @App Content -->
-  <!-- =================================================== -->
-  <div class="blurra">
-
-    <!-- #sidebar ============================ -->
-    <nw-sidebar></nw-sidebar>
-    <!-- #Main ============================ -->
-    <div class="page-container">
-      <!-- #topbar ============================ -->
-      <nw-topbar></nw-topbar>
       <!-- ### $App Screen Content ### -->
         <div ng-view></div>
-        <nw-footer></nw-footer>
-    </div>
-
-
-    <!-- ### $App Screen Footer ### -->
-    
-  
-  </div>
-
-  <!-- ### $App Login Modal ### -->
-  <nw-login></nw-login>
-
 </body>
 </html>
